+++ app/Http/Livewire/Statistics/SeriesPerPublisher.php
@@ -2,7 +2,8 @@
 
 namespace App\Http\Livewire\Statistics;
 
-use App\Constants\SeriesStatus;
+use App\Constants\VolumeStatus;
+use App\Models\Series;
 use Illuminate\Support\Facades\DB;
 use Livewire\Component;
 
@@ -12,16 +13,17 @@ class SeriesPerPublisher extends Component
 
     public function render()
     {
-        $this->seriesByPublisherStatistics = DB::table('series')
-                                                 ->join('publishers', 'series.publisher_id', '=', 'publishers.id')
-                                                 ->where('series.status', '<>', SeriesStatus::CANCELED)
-                                                 ->select('publishers.name as publisher', DB::raw('count(*) as total'))
-                                                 ->groupBy('publishers.name')
-                                                 ->orderByDesc('total')
-                                                 ->orderBy('publishers.name')
-                                                 ->get()
-                                                 ->pluck('total', 'publisher')
-                                                 ->toArray();
+        $this->seriesByPublisherStatistics = Series::join('publishers', 'series.publisher_id', '=', 'publishers.id')
+                                                   ->whereHas('volumes', function ($query): void {
+                                                       $query->WhereIn('status', [VolumeStatus::ORDERED, VolumeStatus::SHIPPED, VolumeStatus::DELIVERED, VolumeStatus::READ]);
+                                                   })
+                                                   ->select('publishers.name as publisher', DB::raw('count(*) as total'))
+                                                   ->groupBy('publishers.name')
+                                                   ->orderByDesc('total')
+                                                   ->orderBy('publishers.name')
+                                                   ->get()
+                                                   ->pluck('total', 'publisher')
+                                                   ->toArray();
 
         return view('livewire.statistics.series-per-publisher');
     }